+++ docker/box.go
@@ -18,8 +18,10 @@ import (
 	"fmt"
 	"io"
 	"io/ioutil"
+	"math"
 	"net/url"
 	"os"
+	"strconv"
 	"strings"
 
 	"github.com/fsouza/go-dockerclient"
@@ -52,6 +54,7 @@ type DockerBox struct {
 	entrypoint      string
 	image           *docker.Image
 	volumes         []string
+	dockerEnvVar    []string
 }
 
 // NewDockerBox from a name and other references
@@ -59,7 +62,7 @@ func NewDockerBox(boxConfig *core.BoxConfig, options *core.PipelineOptions, dock
 	name := boxConfig.ID
 
 	if strings.Contains(name, "@") {
-		return nil, fmt.Errorf("Invalid box name, '@' is not allowed in docker repositories.")
+		return nil, fmt.Errorf("Invalid box name, '@' is not allowed in docker repositories")
 	}
 
 	parts := strings.Split(name, ":")
@@ -117,25 +120,6 @@ func NewDockerBox(boxConfig *core.BoxConfig, options *core.PipelineOptions, dock
 		entrypoint:      entrypoint,
 		volumes:         []string{},
 	}, nil
-}
-
-func (b *DockerBox) links() []string {
-	serviceLinks := []string{}
-
-	for _, service := range b.services {
-		serviceLinks = append(serviceLinks, service.Link())
-	}
-	b.logger.Debugln("Creating links:", serviceLinks)
-	return serviceLinks
-}
-
-// Link gives us the parameter to Docker to link to this box
-func (b *DockerBox) Link() string {
-	name := b.config.Name
-	if name == "" {
-		name = b.ShortName
-	}
-	return fmt.Sprintf("%s:%s", b.container.Name, name)
 }
 
 // GetName gets the box name
@@ -207,19 +191,24 @@ func (b *DockerBox) binds(env *util.Environment) ([]string, error) {
 
 // RunServices runs the services associated with this box
 func (b *DockerBox) RunServices(ctx context.Context, env *util.Environment) error {
-	links := []string{}
-
+	linkedEnvVars := []string{}
 	// TODO(termie): terrible hack, sorry world
 	ctxWithServiceCount := context.WithValue(ctx, "ServiceCount", len(b.services))
 
 	for _, service := range b.services {
 		b.logger.Debugln("Startinq service:", service.GetName())
-		_, err := service.Run(ctxWithServiceCount, env, links)
+		_, err := service.Run(ctxWithServiceCount, env, linkedEnvVars)
 		if err != nil {
 			return err
 		}
-		links = append(links, service.Link())
+		svcEnvVar, err := b.prepareSvcDockerEnvVar(service, env, linkedEnvVars)
+		if err != nil {
+			return err
+		}
+		linkedEnvVars = append(linkedEnvVars, svcEnvVar...)
 	}
+	b.dockerEnvVar = linkedEnvVars
+	b.logger.Debugln(b.dockerEnvVar)
 	return nil
 }
 
@@ -344,7 +333,11 @@ func (b *DockerBox) getContainerName() string {
 
 // Run creates the container and runs it.
 func (b *DockerBox) Run(ctx context.Context, env *util.Environment) (*docker.Container, error) {
-	err := b.RunServices(ctx, env)
+	dockerNetworkName, err := b.GetDockerNetworkName()
+	if err != nil {
+		return nil, err
+	}
+	err = b.RunServices(ctx, env)
 	if err != nil {
 		return nil, err
 	}
@@ -355,6 +348,7 @@ func (b *DockerBox) Run(ctx context.Context, env *util.Environment) (*docker.Con
 
 	// Import the environment
 	myEnv := dockerEnv(b.config.Env, env)
+	myEnv = append(myEnv, b.dockerEnvVar...)
 
 	var entrypoint []string
 	if b.entrypoint != "" {
@@ -389,9 +383,9 @@ func (b *DockerBox) Run(ctx context.Context, env *util.Environment) (*docker.Con
 
 	hostConfig := &docker.HostConfig{
 		Binds:        binds,
-		Links:        b.links(),
 		PortBindings: portBindings(portsToBind),
 		DNS:          b.dockerOptions.DNS,
+		NetworkMode:  dockerNetworkName,
 	}
 
 	conf := &docker.Config{
@@ -449,6 +443,7 @@ func (b *DockerBox) Run(ctx context.Context, env *util.Environment) (*docker.Con
 
 // Clean up the containers
 func (b *DockerBox) Clean() error {
+	defer b.CleanDockerNetwork()
 	containers := []string{}
 	if b.container != nil {
 		containers = append(containers, b.container.ID)
@@ -659,4 +654,73 @@ func (b *DockerBox) ExportImage(options *ExportImageOptions) error {
 	client := b.client
 
 	return client.ExportImage(exportImageOptions)
+}
+
+// Prepares and return DockerEnvironment variables list.
+// For each service In case of docker links, docker creates some environment variables and inject them to box container.
+// Since docker links is replaced by docker network, these environment variables needs to be created manually.
+// Below environment variables created and injected to box container.
+// 01) <container name>_PORT_<port>_<protocol>_ADDR  - variable contains the IP Address.
+// 02) <container name>_PORT_<port>_<protocol>_PORT - variable contains just the port number.
+// 03) <container name>_PORT_<port>_<protocol>_PROTO - variable contains just the protocol.
+// 04) <container name>_ENV_<name> - Docker also exposes each Docker originated environment variable from the source container as an environment variable in the target.
+// 05) <container name>_PORT - variable contains the URL of the source container’s first exposed port. The ‘first’ port is defined as the exposed port with the lowest number.
+// 06) <container name>_NAME - variable is set for each service specified in wercker.yml.
+func (b *DockerBox) prepareSvcDockerEnvVar(service core.ServiceBox, env *util.Environment, linkedEnvVars []string) ([]string, error) {
+	serviceEnv := []string{}
+	client := b.client
+	serviceName := strings.Replace(service.GetServiceAlias(), "-", "_", -1)
+	if containerID := service.GetID(); containerID != "" {
+		container, err := client.InspectContainer(containerID)
+		if err != nil {
+			b.logger.Error("Error while inspecting container", err)
+			return nil, err
+		}
+		ns := container.NetworkSettings
+		var serviceIPAddress string
+		for _, v := range ns.Networks {
+			serviceIPAddress = v.IPAddress
+			break
+		}
+		serviceEnv = append(serviceEnv, fmt.Sprintf("%s_NAME=/%s/%s", strings.ToUpper(serviceName), b.getContainerName(), serviceName))
+		lowestPort := math.MaxInt32
+		var protLowestPort string
+		for k := range container.Config.ExposedPorts {
+			exposedPort := strings.Split(string(k), "/") //exposedPort[0]=portNum and exposedPort[1]=protocal(tcp/udp)
+			x, err := strconv.Atoi(exposedPort[0])
+			if err != nil {
+				b.logger.Error("Unable to convert string port to integer", err)
+				return nil, err
+			}
+			if lowestPort > x {
+				lowestPort = x
+				protLowestPort = exposedPort[1]
+			}
+			dockerEnvPrefix := fmt.Sprintf("%s_PORT_%s_%s", strings.ToUpper(serviceName), exposedPort[0], strings.ToUpper(exposedPort[1]))
+			serviceEnv = append(serviceEnv, fmt.Sprintf("%s=%s://%s:%s", dockerEnvPrefix, exposedPort[1], serviceIPAddress, exposedPort[0]))
+			serviceEnv = append(serviceEnv, fmt.Sprintf("%s_ADDR=%s", dockerEnvPrefix, serviceIPAddress))
+			serviceEnv = append(serviceEnv, fmt.Sprintf("%s_PORT=%s", dockerEnvPrefix, exposedPort[0]))
+			serviceEnv = append(serviceEnv, fmt.Sprintf("%s_PROTO=%s", dockerEnvPrefix, exposedPort[1]))
+		}
+		if protLowestPort != "" {
+			serviceEnv = append(serviceEnv, fmt.Sprintf("%s_PORT=%s://%s:%s", strings.ToUpper(serviceName), protLowestPort, serviceIPAddress, strconv.Itoa(lowestPort)))
+		}
+		for _, envVar := range container.Config.Env {
+			if contains(linkedEnvVars, envVar) == false {
+				serviceEnv = append(serviceEnv, fmt.Sprintf("%s_ENV_%s", strings.ToUpper(serviceName), envVar))
+			}
+		}
+	}
+	b.logger.Debug("Exposed Service Evnironment variables", serviceEnv)
+	return serviceEnv, nil
+}
+
+// Returns true if envVar exist in envVarList.
+func contains(evnVarList []string, envVar string) bool {
+	for _, tmpVar := range evnVarList {
+		if tmpVar == envVar {
+			return true
+		}
+	}
+	return false
 }