+++ docker/service.go
@@ -1,4 +1,4 @@
-//   Copyright 2016 Wercker Holding BV
+//   Copyright © 2016, 2018, Oracle and/or its affiliates.  All rights reserved.
 //
 //   Licensed under the Apache License, Version 2.0 (the "License");
 //   you may not use this file except in compliance with the License.
@@ -106,7 +106,7 @@ func (b *InternalServiceBox) getContainerName() string {
 }
 
 // Run executes the service
-func (b *InternalServiceBox) Run(ctx context.Context, env *util.Environment, links []string) (*docker.Container, error) {
+func (b *InternalServiceBox) Run(ctx context.Context, env *util.Environment, envVars []string) (*docker.Container, error) {
 	e, err := core.EmitterFromContext(ctx)
 	if err != nil {
 		return nil, err
@@ -120,6 +120,7 @@ func (b *InternalServiceBox) Run(ctx context.Context, env *util.Environment, lin
 
 	// Import the environment and command
 	myEnv := dockerEnv(b.config.Env, env)
+	myEnv = append(myEnv, envVars...)
 
 	origEntrypoint := b.image.Config.Entrypoint
 	origCmd := b.image.Config.Cmd
@@ -174,10 +175,15 @@ func (b *InternalServiceBox) Run(ctx context.Context, env *util.Environment, lin
 		portsToBind = b.config.Ports
 	}
 
+	networkName, err := b.GetDockerNetworkName()
+	if err != nil {
+		return nil, err
+	}
+
 	hostConfig := &docker.HostConfig{
 		DNS:          b.dockerOptions.DNS,
 		PortBindings: portBindings(portsToBind),
-		Links:        links,
+		NetworkMode:  networkName,
 	}
 
 	if len(binds) > 0 {
@@ -210,11 +216,20 @@ func (b *InternalServiceBox) Run(ctx context.Context, env *util.Environment, lin
 		conf.MemorySwap = swap
 	}
 
+	endpointConfig := &docker.EndpointConfig{
+		Aliases: []string{b.GetServiceAlias()},
+	}
+	endpointConfigMap := make(map[string]*docker.EndpointConfig)
+	endpointConfigMap[networkName] = endpointConfig
+
 	container, err := client.CreateContainer(
 		docker.CreateContainerOptions{
 			Name:       b.getContainerName(),
 			Config:     conf,
 			HostConfig: hostConfig,
+			NetworkingConfig: &docker.NetworkingConfig{
+				EndpointsConfig: endpointConfigMap,
+			},
 		})
 
 	if err != nil {
@@ -270,6 +285,14 @@ func (b *InternalServiceBox) Run(ctx context.Context, env *util.Environment, lin
 			})
 		}
 	}()
-
 	return container, nil
+}
+
+// GetServiceAlias returns service alias for the service.
+func (b *InternalServiceBox) GetServiceAlias() string {
+	name := b.config.Name
+	if name == "" {
+		name = b.ShortName
+	}
+	return name
 }